create import data page and move functionality from calendar page to export data page
