+++ pages/index.js
@@ -9,7 +9,6 @@ import Calender from 'react-calendar';
 import 'react-calendar/dist/Calendar.css';
 import Snackbar from '../shared/components/Snackbar/Snackbar';
 
-import { addPreviousFoodBlockToLocalStorage, importFoodBlocks } from '../shared/food/food';
 
 export default function Home() {
     const router = useRouter();
@@ -38,68 +37,12 @@ export default function Home() {
     }, []);
 
     const [ snackbars, setSnackbars ] = useState({
-        add: {
-            snackbar: null,
-            timeout: null
-        },
         clear: {
             snackbar: null,
             timeout: null
         }
     });
 
-
-
-    const createAddSnackbar = (message) => {
-        let cloneSnackbars = JSON.parse(JSON.stringify(snackbars));
-        if (cloneSnackbars.add.timeout) {
-            clearTimeout(snackbars.add.timeout);
-        }
-
-        cloneSnackbars.add.snackbar = {
-            message: message,
-            className: 'snackbar-pink'
-        }
-
-        let snackbarTimeout = setTimeout(() => {
-            let cloneSnackbars = JSON.parse(JSON.stringify(snackbars));
-            cloneSnackbars.add.snackbar = null;
-            cloneSnackbars.add.timeout = null;
-            setSnackbars(previousSnackbars => { 
-                return {
-                    ...previousSnackbars,
-                    add: cloneSnackbars.add
-                }
-            });
-        }, 5000);
-
-        cloneSnackbars.add.timeout = snackbarTimeout;
-        setSnackbars(cloneSnackbars);
-    }
-
-    const importData = (event) => {
-        createAddSnackbar('started to import data');
-
-        const reader = new FileReader();
-        reader.onload = (e) => {
-            const importedData = JSON.parse(e.target.result)
-            const previousFoodBlocks = JSON.parse(importedData.previousFoodBlocks);
-
-            previousFoodBlocks.forEach(previousFoodBlock => {
-                addPreviousFoodBlockToLocalStorage({ foodBlock: previousFoodBlock })
-            });
-
-            const foodBlocks = JSON.parse(importedData.foodBlocks)
-            importFoodBlocks(foodBlocks);
-            createAddSnackbar('all data imported');
-        }
-
-        reader.readAsText(event.target.files[0]);
-    }
-
-    const setValueToNull = event => { 
-        event.target.value = null;
-    }
 
     const clearLocalStorage = () => {
         let cloneSnackbars = JSON.parse(JSON.stringify(snackbars));
@@ -158,30 +101,12 @@ export default function Home() {
                 className="m-center" />
 
             <div className="data-buttons">
-
-                <label
-                    htmlFor="import-data"
-                    className="button button-pink cursor-pointer">import data</label>
-                <input
-                    type="file"
-                    id="import-data"
-                    accept=".json"
-                    onChange={ importData }
-                    onClick={ setValueToNull }
-                    className="hidden" />
-
                 <button
                     onClick={ clearLocalStorage }
                     className="button button-red flex-grow-1">clear data</button>
             </div>
 
-
-
             <div className="snackbars-container">
-                { snackbars.add.snackbar ? (
-                    <Snackbar message={ snackbars.add.snackbar.message } className={ snackbars.add.snackbar.className } />
-                ) : (null) }
-
                 { snackbars.clear.snackbar ? (
                     <Snackbar message={ snackbars.clear.snackbar.message } className={ snackbars.clear.snackbar.className } />
                 ) : (null) }
